moving js/css out of static
