+++ src/js/timingDiagram.js
