+++ src/webpack-deps.js
@@ -5,8 +5,8 @@ require('jquery-ui/ui/widgets/draggable');
 exports.Bloodhound = require('corejs-typeahead');
 exports.CodeMirror = require('codemirror');
 exports.marked = require('marked');
-require('./static/assets/js/timingDiagram');
-require('./static/assets/js/igvPatch');
+require('./js/timingDiagram');
+require('./js/igvPatch');
 require('any-resize-event');
 
 require('jquery-ui/themes/base/core.css');
@@ -15,5 +15,5 @@ require('jquery-ui/themes/base/draggable.css');
 require('codemirror/lib/codemirror.css');
 require('github-markdown-css/github-markdown.css');
 require('font-awesome/css/font-awesome.css');
-require('./static/assets/css/markdown.css');
-require('./static/assets/css/twitter-typeahead.css');
+require('./css/markdown.css');
+require('./css/twitter-typeahead.css');
